Default to using jitter, use package-level RNG.
This change defaults to enabling jitter via the standard `New`
function; there is a `NewWithoutJitter` function provided as well.

This change also causes backoff to use a package-level RNG. The other two
approaches were examined and it was decided that the package-level is
the right approach:

+ The global approach is undesirable as packages should not affect
  global state.
+ Placing the RNG in each Backoff structure is undesirable as it adds
  additional complexity and overhead that isn't needed.

The package will panic if the RNG cannot be initialise; given that the
most common usage scenario is for processes under some form of
supervision (e.g. systemd), the panic here is desirable.

Finally, Backoffs are no longer configurable after creation. It is
cleaner (and simpler) to create a new Backoff if new parameters are
chosen.